fix: rename node to nomos-node-mod-da
