+++ nodes/nomos-node-mod-da/config.yaml
