+++ nodes/nomos-node-mod-da/node/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 edition = "2021"
 license = { workspace = true }
-name    = "nomos-node-mod"
+name    = "nomos-node-mod-da"
 version = "0.1.0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
