+++ nodes/nomos-node-mod-da/node/icons/512x512.png
