+++ nodes/nomos-node-mod-da/node/icons/icon.ico
