+++ nomos-da/network/core/src/swarm/common/mod.rs
@@ -2,6 +2,6 @@ pub mod balancer;
 pub mod handlers;
 pub mod monitor;
 pub mod policy;
-pub mod mutated_quic_transport;
+pub mod mutated_transport;
 
 pub use crate::protocols::replication::behaviour::ReplicationConfig;
